% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\docType{data}
\name{area}
\alias{area}
\title{Area of scallop beds}
\format{An object of class \code{data.frame} with 21 rows and 3 columns.}
\usage{
data(area)
}
\description{
Area of scallop beds
}
\details{
Data frame containing abbreviated names, area (squared nautical miles), and number of stations (i.e. grids) for each scallop bed surveyed. Ensure data is current.
}
\examples{
data(area)
}
\references{
See current Statewide Weathervane Scallop survey ROP
}
\keyword{datasets}
