% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_dist_area_names.R
\docType{data}
\name{bed_dist_area_names}
\alias{bed_dist_area_names}
\title{Scallop bed, district and area names.}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 51 rows and 4 columns.}
\usage{
data(bed_dist_area_names)
}
\description{
Scallop bed, district and area names.
}
\details{
Data frame containing bed code, district, and registration area for each bed. Ensure data is current.
}
\examples{
data(bed_dist_area_names)
}
\references{
See current Statewide Weathervane Scallop survey ROP
}
\keyword{datasets}
