% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_catch.R
\name{clean_catch}
\alias{clean_catch}
\title{Title}
\usage{
clean_catch(catch_data, tows, YEAR)
}
\arguments{
\item{YEAR}{}
}
\value{
scal_catch data.frame
}
\description{
Title
}
\examples{

# year of analysis (from "global" head matter in analysis script)
# YEAR <- 2019

scal_catch <- clean_catch(catch_data, tows, YEAR)
}
