% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tow.R
\name{clean_tow}
\alias{clean_tow}
\title{clean_tow}
\usage{
clean_tow(events_data)
}
\arguments{
\item{events_data}{A data.frame of haul events from the weathervane scallop survey}
}
\value{
tows data.frame
}
\description{
clean_tow
}
\examples{

# Must have a defined analysis year
# YEAR <- 2019

tows <- clean_tow(events_data)

}
