% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mw_sh.R
\name{plot_mw_sh}
\alias{plot_mw_sh}
\title{plot_mw_sh}
\usage{
plot_mw_sh(scal_awl, tows, Districts, YEAR)
}
\arguments{
\item{YEAR}{}
}
\description{
Plot scallop meat weight/shell height relationships, by bed.
Saves output to the appropriate figs folder.
}
\examples{
plot_mw_sh(scal_awl, tows, Districts, YEAR)
}
