% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_blister.R
\name{tbl_blister}
\alias{tbl_blister}
\title{tbl_blister}
\usage{
tbl_blister(scal_awl, tows)
}
\arguments{
\item{tows}{}
}
\value{
a .csv table of scallop blister percentage by bed
}
\description{
tbl_blister
}
\examples{
tbl_blister(scal_awl, tows)
}
