% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_catch.R
\name{tbl_catch}
\alias{tbl_catch}
\title{Catch Summary Table}
\usage{
tbl_catch(scal_catch, beds)
}
\arguments{
\item{scal_catch}{Output of `clean_catch()`. See function help file.}

\item{beds}{Output of `clean_bed()`. See function help file.}
}
\value{
Tibble and .csv file
}
\description{
Catch summary by bed
}
\details{
Number of stations and tows for surveyed beds in the 2017 statewide scallop dredge survey with total scallop catches, average scallop densities and corresponding CVs by scallop size class.
}
\examples{
tbl_catch <- tbl_catch(scal_catch, beds)
}
