% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_clap.R
\name{tbl_clap}
\alias{tbl_clap}
\title{Clapper Summary Table}
\usage{
tbl_clap(scal_catch, scal_awl)
}
\arguments{
\item{scal_catch}{Output of `clean_catch()`. See function help file.}

\item{scal_awl}{Output of `clean_awl()`. See function help file.}
}
\value{
Tibble and .csv file
}
\description{
Create a table on prevalence of mud blisters
}
\details{
Bed percentages of clappers and weak meats from survey data. Meat condition was assessed only for subsampled large scallops. N denotes the sample size.
}
\examples{
tbl_clap(scal_catch, scal_awl)
}
