% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_gonad.R
\name{tbl_gonad}
\alias{tbl_gonad}
\title{Gonad Status Summary Table}
\usage{
tbl_gonad(scal_awl, tows)
}
\arguments{
\item{scal_awl}{Output of `clean_awl()`. See function help file.}

\item{tows}{Data frame or tibble containing the fields "tow_id", "Bed", and "area_swept" (case sensitive), pertaining to tows made during the scallop survey year in question. See output of clean_tow().}
}
\value{
Tibble and .csv file
}
\description{
Observed gonad status by bed.
}
\details{
Observed gonad status by bed. Values are percent of sampled scallops ≥100 mm. N denotes the sample size.
}
\examples{
tbl_gonad(scal_awl, tows)
}
