% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_sex.R
\name{tbl_sex}
\alias{tbl_sex}
\title{Sex Summary Table}
\usage{
tbl_sex(scal_awl, tows)
}
\arguments{
\item{scal_awl}{Output of `clean_awl()`. See function help file.}

\item{tows}{Data frame or tibble containing the fields "tow_id", "Bed", and "area_swept" (case sensitive), pertaining to tows made during the scallop survey year in question. See output of clean_tow().}
}
\value{
Tibble and .csv file
}
\description{
Proportion of large scallops by sex.
}
\details{
Observed sex ratios (percent of scallops ≥100 mm). N denotes the sample size.
}
\examples{
tbl_sex(scal_awl, tows)
}
