% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_worm.R
\name{tbl_worm}
\alias{tbl_worm}
\title{Worms Summary Table}
\usage{
tbl_worm(scal_awl, tows)
}
\arguments{
\item{scal_awl}{Output of `clean_awl()`. See function help file.}

\item{tows}{Data frame or tibble containing the fields "tow_id", "Bed", and "area_swept" (case sensitive), pertaining to tows made during the scallop survey year in question. See output of clean_tow().}
}
\value{
Tibble and .csv file
}
\description{
Create a table on prevalence of boring worms.
}
\details{
Area of scallop shells ≥100 mm with evidence of boring worms, by bed. N denotes the sample size.
}
\examples{
tbl_worm(scal_awl, tows)
}
