% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_awl.R
\name{clean_awl}
\alias{clean_awl}
\title{clean_awl}
\usage{
clean_awl(awl_data, tows)
}
\arguments{
\item{tows}{}
}
\value{
clean_awl filters the awl data for successful tows,  changes units to pounds and saves a .csv dataframe to the output folder
}
\description{
clean_awl
}
\examples{

}
