% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meat_wt_est.R
\name{meat_wt_est}
\alias{meat_wt_est}
\title{meat_wt_est}
\usage{
meat_wt_est(scal_awl, scal_catch, beds, Q, boot = TRUE)
}
\arguments{
\item{boot}{}
}
\value{
returns an estimate of meat weight with  bootstrap confidence intervals
}
\description{
meat_wt_est
}
\examples{
# Must have a defined analysis year that is assigned in the "Globals" section
# Year <- 2019
meat_wt_est(scal_awl, scal_catch, beds, Q, boot = TRUE)

}
