% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_rate.R
\name{plot_catch_rate}
\alias{plot_catch_rate}
\title{plot_catch_rate}
\usage{
plot_catch_rate(scal_catch, tows, YEAR)
}
\arguments{
\item{YEAR}{}
}
\description{
Plot survey catch rates for large and small scallops by bed.
Saves output to the appropriate figs folder.
}
\examples{
plot_catch_rate(scal_catch, tows, YEAR)
}
