% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mw_rw.R
\name{plot_mw_rw}
\alias{plot_mw_rw}
\title{plot_mw_rw}
\usage{
plot_mw_rw(scal_awl, tows, Districts, YEAR)
}
\arguments{
\item{YEAR}{}
}
\description{
Plot scallop meat weight/round weight relationships, by bed.
Saves output to the appropriate figs folder.
}
\examples{
plot_mw_rw(scal_awl, tows, Districts, YEAR)
}
