% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scal_est.R
\name{plot_scal_est}
\alias{plot_scal_est}
\title{plot_scal_est}
\usage{
plot_scal_est(scal_est, abundance = TRUE, YEAR)
}
\arguments{
\item{YEAR}{}
}
\description{
Plot scallop abundance (abundance = TRUE) or biomass (abundance = FALSE) estimates by bed.
Saves output to the appropriate figs folder.
}
\examples{
plot_scal_est(scal_est, abundance = TRUE, YEAR)
plot_scal_est(scal_est, abundance = FALSE, YEAR)
}
