% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_size_dist.R
\name{plot_size_dist}
\alias{plot_size_dist}
\title{plot_size_dist}
\usage{
plot_size_dist(scal_awl, scal_catch, tows, YEAR)
}
\arguments{
\item{YEAR}{}
}
\description{
Plot scallop size distributions weighted by sample sizes, by bed.
Saves output to the appropriate figs folder.
}
\examples{
plot_size_dist(scal_awl, scal_catch, tows, YEAR)
}
